Added prototype Derker
Pushing to my branch in a min. Currently he follows the player and teleports to a random spot on the xz plane within 30 units of the player if the player is more than 40 units from him. When he collides with the player he teleports to 0,0,0 and waits 20s before teleporting again.
I gave him a blue spotlight and a fog cloud particle effect so he's easily visible to the player